- added MsgGameTime, to provide a common time to all clients.   it's currently only used for visual information, and is pretty   weak as far as time sync'ing algos go. - allow '-public' in the options block for Thumper
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@11694 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ include/DynamicColor.h
@@ -60,7 +60,7 @@ class DynamicColor {
     void addClampDown(int channel, const float clampDown[3]);
 
     void finalize();
-    void update(float time);
+    void update(double time);
 
     bool canHaveAlpha() const;
     const float* getColor() const;
